move part of SearchableComponent logic into a Trait (task #2207)
