+++ src/Controller/Component/SearchableComponent.php
@@ -8,12 +8,15 @@ use Cake\ORM\Table;
 use Cake\ORM\TableRegistry;
 use \FileSystemIterator;
 use \RuntimeException;
+use Search\Controller\Traits\SearchableTrait;
 
 /**
  * Searchable component
  */
 class SearchableComponent extends Component
 {
+    use SearchableTrait;
+
     /**
      * Default configuration.
      *
@@ -100,32 +103,6 @@ class SearchableComponent extends Component
         }
 
         return array_filter($tables);
-    }
-
-    /**
-     * Returns true if table is searchable, false otherwise.
-     *
-     * @param  \Cake\ORM\Table|string $table Table object or name.
-     * @return bool
-     */
-    public function isSearchable($table)
-    {
-        $result = false;
-        /*
-        get Table instance
-         */
-        if (is_string($table)) {
-            $table = TableRegistry::get($table);
-        }
-
-        /*
-        check if is searchable
-         */
-        if (method_exists($table, 'isSearchable') && is_callable([$table, 'isSearchable'])) {
-            $result = $table->isSearchable();
-        }
-
-        return $result;
     }
 
     /**
